+++ src/main/ru/aston/zhvansky_sv/main_task/domain/Medicine/Drops.java
@@ -1,5 +1,6 @@
 package domain.Medicine;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
@@ -19,7 +20,7 @@ public class Drops extends Medicine {
      * @param activeSubstance The active substance in the drops.
      * @param volume          The volume of the drops in milliliters (ml).
      */
-    public Drops(String name, Double price, Boolean needRecipe, String activeSubstance, Integer volume) {
+    public Drops(String name, BigDecimal price, Boolean needRecipe, String activeSubstance, Integer volume) {
         super(name, price, needRecipe, activeSubstance);
         this.volume = volume;
     }
